Add pycore_atexit.h.
